REF: nanpercentile -> array_algos.quantile
